# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AutoPeace API — AI-powered conflict forecasting and peace research
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /proposals:
    get:
      operationId: listProposals
      tags: [proposals]
      summary: List real-world proposals (US plan, Iran counterproposal, etc.)
      responses:
        "200":
          description: List of proposals
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Proposal"
                required: [data]
    post:
      operationId: createProposal
      tags: [proposals]
      summary: Create a new real-world proposal (admin only)
      security:
        - adminKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProposalRequest"
      responses:
        "201":
          description: Created proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
        "401":
          description: Unauthorized
...
  /proposals/{id}:
    get:
      operationId: getProposal
      tags: [proposals]
      summary: Get a single proposal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Proposal detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
        "404":
          description: Not found
  /admin/proposals/{id}/evaluate:
    post:
      operationId: evaluateProposal
      tags: [proposals, admin]
      summary: Run AI evaluation pipeline on a real-world proposal (admin only)
      description: Runs stakeholder evaluation, judge scoring, and what-would-it-take analysis on the proposal and persists results.
      security:
        - adminKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Evaluated proposal
          content:
            application/json:
              schema:
                type: object
                properties:
                  proposal:
                    $ref: "#/components/schemas/Proposal"
                required: [proposal]
        "401":
          description: Unauthorized
        "404":
          description: Proposal not found
...
  /proposals/submit:
    post:
      operationId: submitPublicProposal
      tags: [proposals, community]
      summary: Submit a real-world proposal for community review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                submitterName: { type: string }
                sourceUrl: { type: string }
                sourceName: { type: string }
                summary: { type: string }
                terms:
                  type: object
                  additionalProperties: true
              required: [sourceUrl, sourceName, summary, terms]
      responses:
        "200":
          description: Submission received
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  message: { type: string }
                required: [id, message]
        "400":
          description: Invalid input
  /admin/proposals/queue:
    get:
      operationId: getProposalQueue
      tags: [admin, proposals]
      summary: Get the admin proposal review queue
      security:
        - adminKey: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
            default: pending
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Proposal queue
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        submitterName: { type: string }
                        sourceUrl: { type: string }
                        sourceName: { type: string }
                        summary: { type: string }
                        status: { type: string }
                        submittedAt: { type: string, format: date-time }
                      required: [id, submitterName, sourceUrl, sourceName, summary, status, submittedAt]
                  total: { type: integer }
                required: [data, total]
        "401":
          description: Unauthorized
  /admin/proposals/queue/{id}:
    patch:
      operationId: reviewProposalSubmission
      tags: [admin, proposals]
      summary: Approve or reject a queued proposal submission
      security:
        - adminKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [approve, reject]
                adminNotes: { type: string }
              required: [action]
      responses:
        "200":
          description: Review applied
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  approvedProposalId:
                    type: string
                    nullable: true
                required: [message]
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "409":
          description: Already reviewed
...
  /admin/proposals/queue/{id}/terms:
    patch:
      operationId: editProposalSubmissionTerms
      tags: [proposals, admin]
      summary: Edit proposal submission summary/terms before approval (admin only)
      security:
        - adminKey: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                summary:
                  type: string
                  description: Updated proposal summary text
                terms:
                  type: object
                  description: Updated deal terms object
      responses:
        "200":
          description: Submission terms updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  id: { type: string }
                required: [message, id]
        "400":
          description: No updates provided
        "401":
          description: Unauthorized
        "404":
          description: Submission not found
        "409":
          description: Cannot edit an already-reviewed submission
...
components:
  securitySchemes:
    adminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
  schemas:
...
    DealTerms:
      type: object
      properties:
        nuclearProtocol:
          type: string
        sanctionsRelief:
          type: string
        hormuzArrangements:
          type: string
        humanitarianProvisions:
          type: string
        verificationMechanism:
          type: string
        timelineYears:
          type: number
        sequencing:
          type: string
        additionalClauses:
          type: array
          items:
            type: string
        stakeholderCommitments:
          type: object
          additionalProperties:
            type: string
          description: Binding commitments from each stakeholder in the grand coalition
    DealScores:
      type: object
      properties:
        feasibility:
          type: number
        coherence:
          type: number
        evidenceGrounding:
          type: number
        domesticSellability:
          type: number
        regionalStability:
          type: number
        implementability:
          type: number
        durability:
          type: number
        composite:
          type: number
    StakeholderVerdict:
      type: object
      properties:
        verdict:
          type: string
          enum: [accept, conditional, reject]
        rationale:
          type: string
        redLineViolations:
          type: array
          items:
            type: string
        conditions:
          type: array
          items:
            type: string
      required:
        - verdict
        - rationale
...
    Proposal:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        source:
          type: string
        submittedBy:
          type: string
        terms:
          $ref: "#/components/schemas/DealTerms"
        scores:
          $ref: "#/components/schemas/DealScores"
          nullable: true
        stakeholderEvaluations:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/StakeholderVerdict"
          nullable: true
        knownResponses:
          type: object
          additionalProperties:
            type: string
          nullable: true
        whatWouldItTake:
          type: array
          items:
            type: object
            additionalProperties: true
          nullable: true
        summary:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - source
        - submittedBy
        - terms
        - summary
        - createdAt
        - updatedAt
...
    CreateProposalRequest:
      type: object
      properties:
        name:
          type: string
        source:
          type: string
        summary:
          type: string
        terms:
          $ref: "#/components/schemas/DealTerms"
      required:
        - name
        - source
        - terms