# lib/api-zod/src/generated/types/dealTerms.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AutoPeace API — AI-powered conflict forecasting and peace research
 * OpenAPI spec version: 0.1.0
 */
import type { DealTermsStakeholderCommitments } from "./dealTermsStakeholderCommitments";

export interface DealTerms {
  nuclearProtocol?: string;
  sanctionsRelief?: string;
  hormuzArrangements?: string;
  humanitarianProvisions?: string;
  verificationMechanism?: string;
  timelineYears?: number;
  sequencing?: string;
  additionalClauses?: string[];
  /** Binding commitments from each stakeholder in the grand coalition */
  stakeholderCommitments?: DealTermsStakeholderCommitments;
}